Style reimbursement list items
