+++ app/helpers/sats-to-btc.js
@@ -1,5 +1,5 @@
 import { helper } from '@ember/component/helper';
 
-export default helper(function satsToBtc(params/*, hash*/) {
-  return params;
+export default helper(function satsToBtc(amount/*, hash*/) {
+  return amount / 100000000;
 });
