+++ app/styles/app.scss
@@ -91,12 +91,13 @@ section {
 @import "sugar";
 @import "item-list";
 
-@import "components/contribution-list";
 @import "components/contribution-details";
+@import "components/contribution-list";
 @import "components/contributor-list";
 @import "components/contributor-profile";
 @import "components/external-account-link";
 @import "components/loading-spinner";
+@import "components/reimbursement-list";
 @import "components/topbar";
 @import "components/topbar-account-panel";
 @import "components/user-avatar";
