+++ app/templates/dashboard/contributions/show.hbs
@@ -19,7 +19,7 @@
     <h3>{{this.model.description}}</h3>
     <p>
       Kind: {{this.model.kind}}
-      <br>Status: {{contribution-status this.model}}
+      <br>Status: {{item-status this.model}}
     </p>
     {{#if this.model.url}}
       <p>
